learning_projects/fraud_detection/worker/requirements.txt
torch
psycopg2-binary
pika
numpy
